search.html 완성

+++ src/main/java/com/swithus/community/club/repository/search/ClubSearchRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface ClubSearchRepository {
-    Page<Object[]> clubSearchPage(Pageable pageable, int region, int sports, String keyword);
+    Page<Object[]> clubSearchPage(Pageable pageable, Long region, Long sports, String keyword);
 }
